refactor(blog): break up gallery.ts god component into smaller Lit components
