+++ apps/blog/src/admin/gallery-utils.ts
@@ -1,0 +1,86 @@
+/**
+ * Shared utility functions for gallery components.
+ */
+
+export const MAX_WIDTH = 2400;
+export const THUMB_WIDTH = 800;
+export const QUALITY = 0.92;
+
+export async function optimizeImage(file: File): Promise<File> {
+  if (file.type === "image/svg+xml") return file;
+  if (file.size < 100 * 1024) return file;
+
+  return new Promise((resolve) => {
+    const img = new Image();
+    img.onload = () => {
+      let { width, height } = img;
+      if (width <= MAX_WIDTH) {
+        resolve(file);
+        return;
+      }
+      height = Math.round(height * (MAX_WIDTH / width));
+      width = MAX_WIDTH;
+      const canvas = document.createElement("canvas");
+      canvas.width = width;
+      canvas.height = height;
+      const ctx = canvas.getContext("2d")!;
+      ctx.imageSmoothingQuality = "high";
+      ctx.drawImage(img, 0, 0, width, height);
+      canvas.toBlob(
+        (blob) => {
+          if (blob) {
+            resolve(new File([blob], file.name, { type: file.type }));
+          } else {
+            resolve(file);
+          }
+        },
+        file.type,
+        QUALITY,
+      );
+    };
+    img.onerror = () => resolve(file);
+    img.src = URL.createObjectURL(file);
+  });
+}
+
+export async function generateThumbnail(file: File): Promise<File> {
+  if (file.type === "image/svg+xml") return file;
+  return new Promise((resolve) => {
+    const img = new Image();
+    img.onload = () => {
+      let { width, height } = img;
+      if (width <= THUMB_WIDTH) {
+        resolve(file);
+        return;
+      }
+      height = Math.round(height * (THUMB_WIDTH / width));
+      width = THUMB_WIDTH;
+      const canvas = document.createElement("canvas");
+      canvas.width = width;
+      canvas.height = height;
+      const ctx = canvas.getContext("2d")!;
+      ctx.imageSmoothingQuality = "high";
+      ctx.drawImage(img, 0, 0, width, height);
+      canvas.toBlob(
+        (blob) => {
+          if (blob) {
+            resolve(new File([blob], file.name.replace(/\.[^.]+$/, ".webp"), { type: "image/webp" }));
+          } else {
+            resolve(file);
+          }
+        },
+        "image/webp",
+        0.8,
+      );
+    };
+    img.onerror = () => resolve(file);
+    img.src = URL.createObjectURL(file);
+  });
+}
+
+export function slugify(text: string): string {
+  return text
+    .toLowerCase()
+    .replace(/[^a-z0-9]+/g, "-")
+    .replace(/^-|-$/g, "");
+}
